pyproject.toml
[project]
name = "coursescrape"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "cryptography>=44.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.27.1",
]
